Product thumbnail image add on admin

+++ routes/web.php
@@ -4,9 +4,11 @@ use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\UserController;
 use App\Http\Controllers\PostController;
 use App\Http\Controllers\ProductController;
-Route::get('/aa', function () {
+
+
+Route::get('/home', function () {
     return view('home');
-});
+})->name('home');
 
 // Route::get('/admin', function () {
 //     return view('admin.index');
@@ -42,20 +44,20 @@ Route::get('posts/', [PostController::class, 'index'])->name('posts');
 
 
 // Products Route
-Route::get('products/', [ProductController::class, 'index'])->name('products');
+Route::get('admin/products/', [ProductController::class, 'index'])->name('products');
 
-Route::get('products/create', [ProductController::class, 'create'])->name('products.create');
+Route::get('admin/products/create', [ProductController::class, 'create'])->name('products.create');
 
-Route::post('products/store', [ProductController::class, 'store'])->name('products.store');
+Route::post('admin/products/store', [ProductController::class, 'store'])->name('products.store');
 
 // Show the edit form for a specific product
-Route::get('products/{product}/edit', [ProductController::class, 'edit'])->name('products.edit');
+Route::get('admin/products/{product}/edit', [ProductController::class, 'edit'])->name('products.edit');
 
 // Handle the update form submission
-Route::put('products/{product}', [ProductController::class, 'update'])->name('products.update');
+Route::put('admin/products/{product}', [ProductController::class, 'update'])->name('products.update');
 
 // Handle product deletion
-Route::delete('products/{product}', [ProductController::class, 'destroy'])->name('products.destroy');
+Route::delete('admin/products/{product}', [ProductController::class, 'destroy'])->name('products.destroy');
 
 // Show a specific product
-Route::get('products/{product}', [ProductController::class, 'show'])->name('products.show');
+Route::get('admin/products/{product}', [ProductController::class, 'show'])->name('products.show');
